requirements.txt
dash
plotly
pandas
umap-learn
